This almost works.
What works:
Show the error message returned by the server for failed input validation.

What doesn't work:
Inputs with errors have a read border instead of black.  This is
done by adding the class `input-validation-error`.

In order to add/remove that class, htmx has to replace the whole
element.  Therefore, there must be a page in the web service that
serves just the input element.  In .NET, that would require putting
each component in its own ViewComponent or PartialPage.  I wouldn't
want to split forms up like that.  A form with 10 inputs would be
20 source files.